refactor: 인수테스트 설정 중복제거
