+++ src/test/java/nextstep/subway/acceptance/AcceptanceTest.java
@@ -1,0 +1,26 @@
+package nextstep.subway.acceptance;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("acceptance")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+    @LocalServerPort
+    int port;
+
+    @Autowired
+    DbCleanup dbCleanup;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+        dbCleanup.execute();
+    }
+
+}
