+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -3,16 +3,10 @@ package nextstep.subway.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.domain.StationRepository;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.net.URI;
 import java.util.Map;
@@ -21,24 +15,7 @@ import static nextstep.subway.acceptance.TestSetupUtils.buildLine;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하촐 노선 관리기능")
-@ActiveProfiles("acceptance")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class LineAcceptanceTest {
-
-    @LocalServerPort
-    int port;
-
-    @Autowired
-    DbCleanup dbCleanup;
-
-    @Autowired
-    StationRepository stationRepository;
-
-    @BeforeEach
-    void setUp() {
-        RestAssured.port = port;
-        dbCleanup.execute();
-    }
+class LineAcceptanceTest extends AcceptanceTest {
 
     /**
      * When 지하철 노선을 생성하면
